chore: Move parts of `fn main into new fn `run`.
According to the documented convention for rust main methods, the
responsibilities of the main method should be

- Calling the command line parsing logic with the argument values
- Setting up any other configuration
- Calling a run function in lib.rs
- Handling the error if run returns an error

With this contribution, `main` sticks to the convention.
